bandwidth: set a custom horizon for fq
The horizon attribute of a fq was introduced in the following kernel commit:

    https://github.com/torvalds/linux/commit/39d010504e6b4485

By default it is set to 10 seconds, and we want to use a lower value, namely 2
seconds. The actual value is defined as bwmap.DefaultDropHorizon and, as of
this commit, is used to configure the t_horizon_drop fields of cilium_throttle
entries, and qdiscs of output devices. At the moment the value is hardcoded.

The horizon attribute of fq is supported starting from kernel v5.8, older
kernels will ignore the setting. To see the current setting on v5.8+ kernels,
use iproute2 v5.14.0+ as follows:

    # tc qdisc show dev DEVICE

Unluckily, this ^ doesn't apply to our case as is, as we are using mq with fq
leaves (by number of device queues). When mq initializes leaves, it does so
using the qdisc_create_dflt function, which doesn't set the ->handle for the
qdisc. The tc_modify_qdisc function doesn't allow to alter qdiscs with zero
handle. So at the moment there is no way to set the horizon parameter for a
mq leaf qdisc.

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>