bpf: allow to enable host routing and endpoint routes simultaneously
We have a long-term goal to enable endpoint routes by default. This patch
removes one of the blockers, namely, that we can not enable endpoint routes
together with optimized routing (bpf host routing).

In order to enforce ingress policy for endpoints, when the endpoint routes are
enabled, we set ForceLocalPolicyEvalAtSource=true. At the moment this is the
only way to enforce the ingress policy when the bpf_redirect_peer() function is
used, because it bypasses the host-side of pod veth pair. [To use the
bpf_redirect() function here doesn't make sense, as this basically reverts all
the optimizations made by the bpf_redirect_peer() function]. ["In future" we
will be able to force the ingress policy at the destination, because the new
"meta" networking device will have a program attached in the pod namespace, and
the bpf_redirect_peer() helper won't bypass it.]

In order to enforce Host Firewall (CCNP), we move a corresponding code section
in bpf_lxc.c one block above.

We still require the cil_to_container program attached to the lxc host part
egress, because we have traffic which is generated by host, so that we can
apply policies to it.

This should work, and policies should be applied once, because the
ipv{4,6}_policy function will eventually execute the redirect_ep() function for
endpoints, which will trigger the bpf_redirect_peer() function, which will
bypass the cil_to_container program, and thus policy will be applied once. This
way policies are applied + we optimize redirects.

This commit is based on the original PR from Daniel Borkmann:

    https://github.com/cilium/cilium/pull/14382

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>