implement network policies using wildcard map
This is a very first attempt to use wildcard map inside cilium.  We now only
have one map lookup in datapath code to match a policy. The userspace is kept
more-or-less the same, however, we map "virtual" policyKeys to "real" keys
on-the-fly (grep RealPolicyKey).

The policy rule is described by the four-tuple (id, dport, proto, direction),
where id and proto may be wildcarded (0 is a magic number which specifies a
wildcard rule), the dport may be a range or a single port or a wildcard
(0-0xffff), and direction is either 0 or 1 (ingress or egress).

In order to use the map we have to hack bpf/bpf_linux.go so that the map is
created with BTF. In spite of the common knowledge that policy map (and some
other maps) is loaded with BTF (from ELF), this is actually not true: they are
created, polupated, and then rechecked with the ELF version. The ELF version is
used if the map has different properties (flags, type, etc.). However, in this
case the policy map is not populated with wildcard entries, so traffic doesn't
flow.

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>