[draft, v0] implement network policies using wildcard map
This is a very first attempt to use wildcard map inside cilium. The hash map is
just replaced with a "hash" version of the wildcard map (all entries are of
type MATCH). The next step is to replace it by proper data structure (and thus
to reduce map lookups from 6 to 1).

In order to use the map we have to hack bpf/bpf_linux.go so that the map is
created with BTF.

In spite of the common knowledge that policy map (and some other maps) is
loaded with BTF (from ELF), this is actually not true: they are created,
polupated, and then rechecked with the ELF version. The ELF version is used if
the map has different properties (flags, type, etc.). However, in this case the
policy map is not populated with wildcard entries, so traffic doesn't flow.

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>