[draft] implement network policies using native wildcard
We now only have one map lookup in datapath code. The userspace is kept
more-or-less the same, however, we map "virtual" policyKeys to "real" keys
on-the-fly (grep RealPolicyKey).

The policy rule is described by the four-tuple (id, dport, proto, direction),
where id and proto may be wildcarded (0 is a magic number which specifies a
wildcard rule), the dport may be a range or a single port or a wildcard
(0-0xffff), and direction is either 0 or 1 (ingress or egress).

TBD: blah-blah-blah better and describe the redesign

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>