[draft] fix ipv6 neighbour discovery by installing neighbour entries
As we are already installing per-endpoint routes, this is simple to install
IPv6 neighbour entries in sync. This fixes the neighbour discovery in case
endpoint routes and host routing are enabled together (in the opposite case we
drop ICMPv6 NA messages from pods as they are destined to link-local addresses,
see also https://github.com/cilium/cilium/issues/23910 for a similar problem).